+++ lib/dart/MemoryState.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:switchboard/dart/globalHelpers.dart';
+import 'package:switchboard/globalHelpers.dart';
 
 class MemoryState {
   static final MemoryState _state = MemoryState._init();
@@ -457,8 +458,9 @@ class MemoryState {
     return js;
   }
 
-  void reset() {
+  Future<void> reset() async {
     fromJson(_defaults.toJson());
+    await clearApplicationFont();
   }
 }