+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - [x] Fixes another bug in `NetworkInterface.requestAltersList` which prevented the Web App specifically from loading the entire alters list for larger systems.
 - [x] Attempt to fix the WebApp text alignment
 - [ ] Implement requested feature (#7), ability to turn off the glow animations
+- [x] Fix #10: Reset default will now clear the custom font.
 
 # 0.3.1+0705260030